styling of AnimalCard completed, functional
