+++ components/Animals/Animals.js
@@ -1,11 +1,10 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { View, StyleSheet, FlatList, ScrollView } from 'react-native';
 
 import AnimalCard from './AnimalCard';
 import animalData from './animalData';
 const Animals = props => {
   return (
-    // <View style={styles.animalsList}>
     <ScrollView>
       <View>
         <FlatList
@@ -17,15 +16,14 @@ const Animals = props => {
         />
       </View>
     </ScrollView>
-    // </View>
   );
 };
 
-const styles = StyleSheet.create({
-  animalsList: {
-    flex: 0,
-    width: '120%'
-  }
-});
+// const styles = StyleSheet.create({
+//   animalsList: {
+//     flex: 0,
+//     width: '120%'
+//   }
+// });
 
 export default Animals;
